Change some int types

+++ src/samplers/mcmc/ahmc/ahmc_metric.jl
@@ -9,14 +9,14 @@ struct DenseEuclideanMetric <: HMCMetric end
 
 
 
-function AHMCMetric(metric::DiagEuclideanMetric, dim::Int64)
+function AHMCMetric(metric::DiagEuclideanMetric, dim::Integer)
     return AdvancedHMC.DiagEuclideanMetric(dim)
 end
 
-function AHMCMetric(metric::UnitEuclideanMetric, dim::Int64)
+function AHMCMetric(metric::UnitEuclideanMetric, dim::Integer)
     return AdvancedHMC.UnitEuclideanMetric(dim)
 end
 
-function AHMCMetric(metric::DenseEuclideanMetric, dim::Int64)
+function AHMCMetric(metric::DenseEuclideanMetric, dim::Integer)
     return AdvancedHMC.DenseEuclideanMetric(dim)
 end
